cli/requirements.txt
rich>=13.0.0
questionary>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
openai>=1.0.0
anthropic>=0.18.0
httpx>=0.25.0
click>=8.1.0
requests>=2.31.0
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0
